% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeroSampleR-package.R
\docType{package}
\name{AeroSampleR-package}
\alias{AeroSampleR}
\alias{AeroSampleR-package}
\title{AeroSampleR: Estimate Aerosol Particle Collection Through Sample Lines}
\description{
Estimate ideal efficiencies of aerosol sampling through sample lines. Functions were developed consistent with the approach described in Hogue, Mark; Thompson, Martha; Farfan, Eduardo; Hadlock, Dennis, (2014), "Hand Calculations for Transport of Radioactive Aerosols through Sampling Systems" Health Phys 106, 5, S78-S87, \doi{10.1097/HP.0000000000000092}.
}
\author{
\strong{Maintainer}: Mark Hogue \email{mark.hogue.chp@gmail.com}

}
\keyword{internal}
