% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeroSampleR-package.R
\name{AeroSampleR}
\alias{AeroSampleR}
\title{Estimate Aerosol Particle Collection Through Sample Lines}
\description{
This package provides a method to estimate sampling
efficiency of sampling systems drawing aerosol particles
through tubing.
}
\details{
Functions were developed consistent with the approach
described in Hogue, Mark; Thompson, Martha; Farfan,
Eduardo; Hadlock, Dennis, (2014), "Hand Calculations for
Transport of Radioactive Aerosols through Sampling Systems"
Health Phys 106, 5, S78-S87, <doi:10.1097/HP.0000000000000092>.

To learn how to use AeroSampleR, start with the vignette:
`browseVignettes(package = "AeroSampleR")`
}
