% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_for_plots.R
\docType{data}
\name{dat_for_plots}
\alias{dat_for_plots}
\title{Data from readme file for use in plot examples}
\format{
A \code{data.frame}
\describe{
\item{D_p}{particle diameter in micrometers}
\item{dens}{probability density}
\item{dist}{either log_norm or discrete}
\item{C_c}{Cunningham slip correction factor}
\item{v_ts}{particle terminal velocity}
\item{Re_p}{Reynold's number for particle}
\item{Stk}{Stokes' number for particle}
\item{eff_probe}{aspiration efficiency for probe}
\item{eff_bend_2}{transport efficiency for the second component, a bend}
\item{eff_tube_3}{transport efficiency for the third component, a straight tube}
}
}
\usage{
dat_for_plots
}
\description{
This data was created by running the readme
script. It is needed for simple plot examples.
}
\keyword{datasets}
