% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cum_plots.R
\name{report_cum_plots}
\alias{report_cum_plots}
\title{report on cumulative transport system efficiency (discrete particle sizes only)}
\usage{
report_cum_plots(df, micron)
}
\arguments{
\item{df}{is the particle data set - after transport analysis by element}

\item{micron}{selects the particle size (aerodynamic mass activity
diameter in micrometers). This must be selected from the original
distribution of particles that were started with the `particle_dist`
function.}
}
\value{
A plot of cumulative transport efficiencies is generated in a plot window
}
\description{
In order to run a report, first produce a model of each individual
element. Start with producing a particle distribution
with the `particle_dist` function, then produce a parameter set with
the `set_params` function. Both of these results must be stored as
per examples described in the help set with each. Next, add elements
in the sample system until all are complete.
}
\examples{
report_cum_plots(dat_for_plots, micron = 10)

}
