% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plots.R
\name{report_plots}
\alias{report_plots}
\title{plots of individual on transport system elements}
\usage{
report_plots(df, dist)
}
\arguments{
\item{df}{is the particle data set - after transport analysis by element}

\item{dist}{selects the distribution for the report. Options are
'discrete' for discrete particle sizes or 'log' for the log-normal
distribution of particles that were started with the `particle_dist`
function.}
}
\value{
A plot of transport efficiencies is generated in a plot window
}
\description{
In order to run a report, first produce a model of each individual
element. Start with producing a particle distribution
with the `particle_dist` function, then produce a parameter set with
the `set_params` function. Both of these results must be stored as
per examples described in the help set with each. Next, add elements
in the sample system until all are complete.
}
\examples{
report_plots(dat_for_plots, dist = 'discrete')

}
