% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_params_2.R
\name{set_params_2}
\alias{set_params_2}
\title{Make a set of parameters that will be used throughout this package.
Carry these parameters forward in the data set with the particle
distribution.
`set_params_1` sets all single parameters.
`set_params_2` adds particle size-dependent parameters to the particle
distribution}
\usage{
set_params_2(df, params)
}
\arguments{
\item{df}{is the particle data set (data frame) established with the
`particle_dist` function}

\item{params}{is the parameter data set for parameters that are not
particle size-dependent}
}
\value{
a data frame starting with the submitted particle distribution
 with additional columns for particle-size-dependent parameters
}
\description{
All parameters are to be in MKS units, except as noted.
}
\examples{
\dontrun{df <- set_params_2(df, params)}

}
