% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bend_eff.R
\name{bend_eff}
\alias{bend_eff}
\title{bend efficiency}
\usage{
bend_eff(df, params, method, bend_angle, bend_radius, elnum)
}
\arguments{
\item{df}{is the particle data set (data frame) established with the
`particle_dist` function}

\item{params}{is the parameter data set for parameters that are not
particle size-dependent}

\item{method}{choice of models: Pui, McFarland, or Zhang}

\item{bend_angle}{bend angle in degrees}

\item{bend_radius}{bend radius in m}

\item{elnum}{element number to provide uniqe column names}
}
\value{
data frame containing original particle distribution with added
data for this element
}
\description{
In order to run this function, first produce a particle distribution
with the `particle_dist` function, then produce a parameter set with
the `set_params` function. Both of these results must be stored as
per examples described in the help set with each.
}
\examples{
\dontrun{df <- bend_eff(df, params, method='Zhang', bend_angle=90,
bend_radius=0.1, elnum=3)}
}
\references{
A. R. McFarland, H. Gong, A. Muyshondt, W. B. Wente, and N. K. Anand
Environmental Science & Technology 1997 31 (12), 3371-3377
<doi:10.1021/es960975c>

Pusheng Zhang, Randy M. Roberts, André Bénard,
Computational guidelines and an empirical model for particle deposition
in curved pipes using an Eulerian-Lagrangian approach,
Journal of Aerosol Science, Volume 53, 2012, Pages 1-20,ISSN 0021-8502,
<doi:10.1016/j.jaerosci.2012.05.007>

David Y. H. Pui, Francisco Romay-Novas & Benjamin Y. H. Liu (1987)
Experimental Study of Particle Deposition in Bends of Circular Cross
Section, Aerosol Science and Technology, 7:3, 301-315,
<doi:10.1080/02786828708959166>
}
