% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_basic.R
\name{report_basic}
\alias{report_basic}
\title{report on transport efficiency}
\usage{
report_basic(df, params, dist)
}
\arguments{
\item{df}{is the particle data set (data frame) established with the
`particle_dist` function}

\item{params}{is the parameter data set for parameters that are not
particle size-dependent}

\item{dist}{selects the distribution for the report. Options are
"discrete" for discrete particle sizes or "log" for the log-normal
distribution of particles that were started with the `particle_dist`
function.}
}
\value{
report of system efficiency
}
\description{
In order to run a report, first produce a model of each individual
element. Start with producing a particle distribution
with the `particle_dist` function, then produce a parameter set with
the `set_params` function. Both of these results must be stored as
per examples described in the help set with each. Next, add elements
in the sample system until all are complete.
}
\examples{
\dontrun{report_eff(df, params, dist = "discrete")}

}
