% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_cum_plots.R
\name{report_cum_plots}
\alias{report_cum_plots}
\title{report on cumulative transport system efficiency}
\usage{
report_cum_plots(df, micron)
}
\arguments{
\item{df}{is the particle data set - after transport analysis by element}

\item{micron}{selects the particle size (aerodynamic mass activity
diameter in micrometers). This must be selected from the original
distribution of particles that were started with the `particle_dist`
function.}
}
\value{
A plot of cumulative transport efficiencies is generated in a plot window
}
\description{
In order to run a report, first produce a model of each individual
element. Start with producing a particle distribution
with the `particle_dist` function, then produce a parameter set with
the `set_params` function. Both of these results must be stored as
per examples described in the help set with each. Next, add elements
in the sample system until all are complete.
}
\examples{
df <- particle_dist(n=10) # toy set for demo - set up particle distribution
params <- set_params_1("D_tube" = 2.54, "Q_lpm" = 100,
"T_C" = 25, "P_kPa" = 101.325) #example system parameters
df <- set_params_2(df, params) #particle size-dependent parameters
df <- probe_eff(df, params, orient = 'h') #probe orientation - horizontal
df <- bend_eff(df, params, method='Zhang', bend_angle=90,
bend_radius=0.1, elnum=3)
df <- tube_eff(df, params, L = 100,
angle_to_horiz = 90, elnum = 3)
report_cum_plots(df, micron = 10)

}
