% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_log_mass.R
\name{report_log_mass}
\alias{report_log_mass}
\title{report relative masses by particle of a log-normal distribution}
\usage{
report_log_mass(df)
}
\arguments{
\item{df}{is the particle data set - after transport analysis by element}
}
\value{
data frame containing mass-based particle fractions in ambient
location and in distribution delivered throught the system.
}
\description{
This function shows the entire table of results by particle diameter.
}
\examples{
\dontrun{log_mass_results <- report_log_mass(df)}

}
